add query analyzer with min and avg similarity
